1. Change copy icon to copy code instead of link
2. Added more parameters to ctor to customize the widget

3. Changed folder/file structure to comform to standard

4. Added copyWith to SyntaxHighlighterStyle

5. Modified example app by removing example.dart and moving everything into main.dart

6. Changed example app to show its own code and made it more verbose
