--- example/lib/example.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Example extends StatelessWidget {
-  const Example({Key key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("An example widget"),
-    );
-  }
-}
