+++ example/pubspec.yaml
@@ -30,4 +30,4 @@ flutter:
   uses-material-design: true
 
   assets:
-    - lib/example.dart
+    - lib/
