+++ example/test/widget_test.dart
@@ -5,15 +5,14 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:example/main.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
+// import 'package:example/main.dart';
+// import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+  // testWidgets('smoke test', (WidgetTester tester) async {
+  //   // Build our app and trigger a frame.
+  //   // await tester.pumpWidget(MyApp());
 
-    expect(find.text('1'), findsNothing);
-  });
+  //   // expect(find.text('1'), findsNothing);
+  // });
 }
